+++ packages/tab/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.2](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@11.0.1...@pluralsight/ps-design-system-tab@11.0.2) (2019-06-04)
+
+
+### Bug Fixes
+
+* **tab:** dont cut off active orange bar ([b970d72](https://github.com/pluralsight/design-system/commit/b970d72))
+* **tab:** prevent overflow buttons from shifting on list focus ([88af4d6](https://github.com/pluralsight/design-system/commit/88af4d6))
+
+
+
+
+
+## [11.0.1](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@11.0.0...@pluralsight/ps-design-system-tab@11.0.1) (2019-06-03)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-tab
+
+
+
+
+
 # [11.0.0](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-tab@10.0.0...@pluralsight/ps-design-system-tab@11.0.0) (2019-05-31)